+++ packages/yew-functional/tests/use_context.rs
@@ -3,7 +3,7 @@ mod common;
 use common::obtain_result_by_id;
 use std::rc::Rc;
 use wasm_bindgen_test::*;
-use yew::{html, Children, Html, Properties};
+use yew::{html, Children, ContextProvider, Html, Properties};
 use yew_functional::{
     use_context, use_effect, use_ref, use_state, FunctionComponent, FunctionProvider,
 };
